fix(its-battistar-angular): :bug: fast-json-stringify was importing node only lib
BREAKING CHANGE: created new lib

closes #69
