+++ libs/shared-utils/eslint.config.js
@@ -1,11 +1,33 @@
+/* eslint-disable unicorn/prefer-module */
+const globals = require('globals');
 const baseConfig = require('../../eslint.config.js');
+const tseslint = require('typescript-eslint');
 
-module.exports = [
+module.exports = tseslint.config(
   ...baseConfig,
+
   {
-    files: ['**/*.ts', '**/*.tsx', '**/*.js', '**/*.jsx'],
-    rules: {},
+    ignores: ['!**/*', 'node_modules', 'dist'],
   },
+
+  {
+    languageOptions: {
+      parserOptions: {
+        project: ['tsconfig.*?.json'],
+        tsconfigRootDir: __dirname,
+      },
+      globals: {
+        ...globals.node,
+        ...globals.es2021,
+        ...globals.browser,
+      },
+    },
+  },
+  // {
+  //   files: ['**/*.ts', '**/*.tsx', '**/*.js', '**/*.jsx'],
+  //   rules: {},
+  //   languageSettings: { parserOptions: { project: ['tsconfig.*?.json'] } },
+  // },
   {
     files: ['**/*.ts', '**/*.tsx'],
     rules: {},
@@ -13,5 +35,5 @@ module.exports = [
   {
     files: ['**/*.js', '**/*.jsx'],
     rules: {},
-  },
-];
+  }
+);
