+++ libs/shared-utils/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@its-battistar/shared-utils",
+  "version": "1.0.0"
+}
