+++ libs/shared-utils/project.json
@@ -5,11 +5,28 @@
   "projectType": "library",
   "tags": [],
   "targets": {
+    "pre-lint": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "tsc --noEmit -p ./tsconfig.lib.json && prettier -c ./src",
+        "cwd": "libs/shared-types"
+      }
+    },
+    "lint": {
+      "dependsOn": ["pre-lint"],
+      "executor": "@nx/eslint:lint"
+    },
     "test": {
       "executor": "@nx/vite:test",
       "outputs": ["{options.reportsDirectory}"],
       "options": {
         "reportsDirectory": "../../coverage/libs/shared-utils"
+      }
+    },
+    "version": {
+      "executor": "@jscutlery/semver:version",
+      "options": {
+        "preset": "conventionalcommits"
       }
     }
   }
