--- libs/shared-utils/src/lib/shared-utils.ts
@@ -1,3 +1,0 @@
-export function sharedUtils(): string {
-  return 'shared-utils';
-}
