+++ libs/shared-utils/src/lib/utils/account.ts
@@ -1,0 +1,12 @@
+import { accountSchema } from '@its-battistar/shared-types';
+import fastJsonStringify from 'fast-json-stringify';
+
+import ajvInstance from './ajv';
+
+/**
+ * Validation and serialization functions for the schemas provided
+ */
+
+export const stringifyAccount = fastJsonStringify(accountSchema);
+
+export const validateAccount = ajvInstance.compile(accountSchema);
