+++ libs/shared-utils/src/lib/utils/todo.ts
@@ -1,0 +1,9 @@
+import { todoSchema, todoSchemaInput } from '@its-battistar/shared-types';
+import fastJsonStringify from 'fast-json-stringify';
+
+import ajvInstance from './ajv';
+
+// Utility functions
+export const stringifyTodo = fastJsonStringify(todoSchema);
+
+export const validateTodo = ajvInstance.compile(todoSchemaInput);
