+++ libs/shared-utils/src/lib/utils/user.ts
@@ -1,0 +1,9 @@
+import { userSchema, userSchemaInput } from '@its-battistar/shared-types';
+import fastJsonStringify from 'fast-json-stringify';
+
+import ajvInstance from './ajv';
+
+// TODO: fix validation for different use cases and data types
+export const stringifyUser = fastJsonStringify(userSchema);
+
+export const validateUser = ajvInstance.compile(userSchemaInput);
