+++ libs/shared-utils/tsconfig.eslint.json
@@ -1,0 +1,4 @@
+{
+  "extends": "./tsconfig.json",
+  "include": ["./**/*.js", "./**/*.ts"]
+}
