+++ libs/shared-utils/vite.config.ts
@@ -1,6 +1,5 @@
-import { defineConfig } from 'vite';
-
 import { nxViteTsPaths } from '@nx/vite/plugins/nx-tsconfig-paths.plugin';
+import { defineConfig } from 'vite';
 
 export default defineConfig({
   root: __dirname,
